+++ Assets/Project/Scripts/MenuSelectScene/MenuSelectDirector.cs
@@ -1,42 +1,21 @@
 ﻿using System.Collections;
 using Project.Scripts.Utils.Definitions;
-using UnityEngine;
+using Project.Scripts.Utils.Patterns;
 using UnityEngine.SceneManagement;
-using UnityEngine.UI;
 
 namespace Project.Scripts.MenuSelectScene
 {
-    public class MenuSelectDirector : MonoBehaviour
+    public class MenuSelectDirector : SingletonObject<MenuSelectDirector>
     {
-        private const string STAGE_SELECT_TOGGLE_NAME = "StageSelect";
-        private const string RECORD_TOGGLE_NAME = "Record";
-        private const string TUTORIAL_TOGGLE_NAME = "Tutorial";
-        private const string CONFIG_TOGGLE_NAME = "Config";
-
-        private string nowScene;
-
-        private GameObject stageSelectToggle;
-
-        private GameObject recordToggle;
-
-        private GameObject tutorialToggle;
-
-        private GameObject configToggle;
+        public string nowScene;
 
         private void Awake()
         {
-            // Toggleの取得
-            stageSelectToggle = GameObject.Find(STAGE_SELECT_TOGGLE_NAME);
-            recordToggle = GameObject.Find(RECORD_TOGGLE_NAME);
-            tutorialToggle = GameObject.Find(TUTORIAL_TOGGLE_NAME);
-            configToggle = GameObject.Find(CONFIG_TOGGLE_NAME);
-            // Toggleのリスナーを設定
-            AddListeners();
             // 初期シーンのロード
             StartCoroutine(AddScene(SceneName.STAGE_SELECT_SCENE));
         }
 
-        private IEnumerator AddScene(string sceneName)
+        public IEnumerator AddScene(string sceneName)
         {
             // シーンをロード
             SceneManager.LoadScene(sceneName, LoadSceneMode.Additive);
@@ -50,41 +29,6 @@ namespace Project.Scripts.MenuSelectScene
             SceneManager.SetActiveScene(scene);
             // シーンの保存
             nowScene = sceneName;
-        }
-
-        private void AddListeners()
-        {
-            stageSelectToggle.GetComponent<Toggle>().onValueChanged.AddListener(delegate {
-                ToggleValueChanged(stageSelectToggle);
-            });
-
-            recordToggle.GetComponent<Toggle>().onValueChanged
-            .AddListener(delegate {
-                ToggleValueChanged(recordToggle);
-            });
-
-            tutorialToggle.GetComponent<Toggle>().onValueChanged
-            .AddListener(delegate {
-                ToggleValueChanged(tutorialToggle);
-            });
-
-            configToggle.GetComponent<Toggle>().onValueChanged
-            .AddListener(delegate {
-                ToggleValueChanged(configToggle);
-            });
-        }
-
-        private void ToggleValueChanged(GameObject toggle)
-        {
-            // ONになった場合のみ処理
-            if (toggle.GetComponent<Toggle>().isOn) {
-                // 現在チェックされている Toggle を OFF にする
-                GameObject.Find(nowScene.Replace("Scene", "")).GetComponent<Toggle>().isOn = false;
-                // 今のシーンをアンロード
-                SceneManager.UnloadSceneAsync(nowScene);
-                // 新しいシーンをロード
-                StartCoroutine(AddScene(toggle.name + "Scene"));
-            }
         }
     }
 }
