--- Assets/Project/Scripts/Utils/MyUiConponents.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 688af07ac03ab4f1d8e5d1a7fd42b02e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
