--- Assets/Project/Scripts/Utils/MyUiConponents/MyToggle.cs
@@ -1,23 +1,0 @@
-﻿using UnityEngine.EventSystems;
-using UnityEngine.UI;
-
-namespace Project.Scripts.Utils.MyUiConponents
-{
-    public class MyToggle : Toggle
-    {
-        /// <summary>
-        /// ToggleGroup の AllowSwitchOff=false を再現
-        /// <see cref="Toggle.OnPointerClick"/>.
-        /// </summary>
-        public override void OnPointerClick(PointerEventData eventData)
-        {
-            if (eventData.button != PointerEventData.InputButton.Left) return;
-
-            if (!IsActive() || !IsInteractable())
-                return;
-
-            // ON 状態のものを OFF にすることは許さない
-            if (!isOn) isOn = true;
-        }
-    }
-}
