--- Assets/Project/Scripts/Utils/MyUiConponents/MyToggle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 974c4e29a30f04b28967506658c9c100
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
